requirements.txt
# 📦 Requirements - StarkAdvisor

# Dependencias principales
Django==5.2.5
djangorestframework==3.15.2
django-cors-headers==4.4.0
Pillow==10.4.0

# PostgreSQL
psycopg2-binary==2.9.10

# Redis para sesiones y caché
redis==5.0.1
django-redis==5.4.0

# Autenticación
djangorestframework-authtoken

# Dependencias automáticas (se instalan con las principales)
asgiref==3.8.1
pytz==2024.1
sqlparse==0.5.1
typing-extensions==4.12.2
